requirements.txt
alembic==1.4.3
altgraph==0.17
aniso8601==9.0.1
appdirs==1.4.4
astroid==2.5.7
attrs==21.2.0
autopep8==1.5.7
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.3
blinker==1.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
cryptography==2.9.2
decorator==4.4.2
distlib==0.3.1
dnspython==2.1.0
docopt==0.6.2
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.4
Flask-Cors==3.0.10
Flask-DebugToolbar==0.11.0
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-PyMongo==2.3.0
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
h11==0.12.0
h2==4.0.0
hpack==4.0.0
hyperframe==6.0.0
idna==2.10
importlib-metadata==4.4.0
ipython==7.21.0
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.3
Js2Py==0.71
jsonschema==3.2.0
kaitaistruct==0.9
lazy-object-proxy==1.6.0
lxml==4.6.3
Mako==1.1.3
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.20.2
opencv-python==4.5.2.52
openpyxl==3.0.6
packaging==20.9
pandas==1.2.3
parso==0.8.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==8.1.0
pipenv==2020.11.15
pipwin==0.5.1
pocketsphinx @ file:///C:/Users/WIN-10/pipwin/pocketsphinx-0.1.15-cp37-cp37m-win_amd64.whl
prompt-toolkit==3.0.16
psycopg2==2.8.6
pyasn1==0.4.8
PyAudio @ file:///C:/Users/WIN-10/pipwin/PyAudio-0.2.11-cp37-cp37m-win_amd64.whl
pycodestyle==2.7.0
pycparser==2.20
pydivert==2.1.0
pydub==0.24.1
pygame==2.0.1
Pygments==2.8.0
pyjsparser==2.7.1
PyJWT==2.1.0
pylint==2.8.2
pymongo==3.11.4
pyOpenSSL==20.0.1
pyparsing==2.4.7
pypiwin32==223
PyPrind==2.11.3
pyrsistent==0.17.3
pySmartDL==1.3.4
PySocks==1.7.1
pytesseract==0.3.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
pywin32==300
pyxll==5.0.9
requests==2.25.1
selenium==3.141.0
selenium-wire==4.2.2
sentry-sdk==0.18.0
six==1.16.0
soupsieve==2.1
SpeechRecognition==3.8.1
SQLAlchemy==1.3.19
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.5
virtualenv==20.4.2
virtualenv-clone==0.5.4
wcwidth==0.2.5
webdriver-manager==3.2.2
Werkzeug==0.16.1
wrapt==1.12.1
wsproto==1.0.0
WTForms==2.3.3
xlrd==2.0.0
xlwings==0.22.2
zipp==3.4.1
